Connected Textboxes to Cells, Added color changing, Fixed CheckWin, named Textboxes and renamed columns
Connected the 2D array of textboxes to the 2D array of cells. I added the color changing functionality and fixed some errors with the CheckWin code as it was looking out of bounds on some occasions during testing. Finally, I renamed the columns and textboxes because it didn't save the last time I did for some reason. It's starting to work like an actual connect 4 game now!